Move GET medication request from Context.js to CatalogMain.js to remove unnecessary requests to server
