+++ src/Context.js
@@ -1,105 +1,16 @@
-import { createContext, useState, useRef, useEffect } from 'react';
+import { createContext, useState } from 'react';
 
 const Context = createContext();
 
 const ContextProvider = (props) => {
-    const [allMedications, setAllMedications] = useState([]);
     const [cartMedications, setCartMedications] = useState([]);
     const [cartItemsId, setCartItemsId] = useState(localStorage.getItem('cart') ?
         JSON.parse(localStorage.getItem('cart')) : []);
-
-    const [availability, setAvailability] = useState(true);
     const [selectedItems, setSelectedItems] = useState([]);
 
-    const refPage = useRef(0);
-    const [allPages, setAllPages] = useState(0);
-    const [totalMed, setTotalMed] = useState(0);
-
     const [droneId, setDroneId] = useState();
-
     const [deliveryDetail, setDeliveryDetail] = useState();
     const [isReadyPostFetch, setIsReadyPostFetch] = useState(false);
-
-    const urlFilters = {
-        available: "status=AVAILABLE",
-        sort: {
-            nameAZ: "sort=name,asc",
-            nameZA: "sort=name,desc",
-            weightUp: "sort=weight,asc",
-            weightDown: "sort=weight,desc",
-        }
-    };
-    const [urlSelect, setUrlSelect] = useState(urlFilters.sort.nameAZ);
-    const href = "http://localhost:8090/medication?size=6&page=";
-
-    const [urlMedication, setUrlMedication] = useState(href + refPage.current + "&" + urlFilters.available + "&" + urlSelect);
-
-    useEffect(() => {
-        let ignore = false;
-        fetch(urlMedication, {
-            method: 'GET',
-            headers: { 'Content-Type': 'application/json' }
-        })
-            .then(result => result.json())
-            .then(data => {
-                if (!ignore) {
-                    setAllPages(data.totalPages);
-                    setTotalMed(data.totalElements);
-                    updateWithCartInfo(data.content);
-                }
-            });
-        return () => {
-            ignore = true;
-        }
-    }, [urlMedication, cartItemsId]);
-
-    const updateWithCartInfo = (medications) => {
-        const updateMedications = medications.map(med => {
-            const isInCart = cartItemsId.find(item => item === med.id);
-            if (isInCart) {
-                return { ...med, status: "UNAVAILABLE" };
-            } else {
-                return { ...med, status: med.status };
-            }
-        });
-        setAllMedications(updateMedications);
-    }
-
-    const toggleAvailable = (value) => {
-        refPage.current = 0;
-        setAvailability(value);
-        setUrlMedication(value
-            ? href + refPage.current + "&" + urlFilters.available + "&" + urlSelect
-            : href + refPage.current + "&" + urlSelect);
-    }
-
-    const selectSort = (value) => {
-        refPage.current = 0;
-        for (let key in urlFilters.sort) {
-            if (value === key) {
-                setUrlMedication(availability
-                    ? href + refPage.current + "&" + urlFilters.available + "&" + urlFilters.sort[value]
-                    : href + refPage.current + "&" + urlFilters.sort[value]);
-                setUrlSelect(urlFilters.sort[value]);
-            }
-        }
-    }
-
-    const currentPage = (value) => {
-        switch (value) {
-            case "left":
-                refPage.current = refPage.current === 0 ? refPage.current = 0 : refPage.current - 1;
-                break;
-            case "right":
-                refPage.current = refPage.current === allPages - 1
-                    ? allPages - 1 : refPage.current + 1;
-                break;
-            default: refPage.current = value;
-        }
-        setUrlMedication(availability
-            ? href + refPage.current + "&" + urlFilters.available + "&" + urlSelect
-            : href + refPage.current + "&" + urlSelect);
-    }
 
     const append = (id) => {
         if (cartItemsId.length === 0) {
@@ -125,19 +36,12 @@ const ContextProvider = (props) => {
     }
 
     const value = {
-        allMedications,
         cartMedications,
         setCartMedications,
         cartItemsId,
+        setCartItemsId,
         selectedItems,
         setSelectedItems,
-        availability,
-        toggleAvailable,
-        selectSort,
-        currentPage,
-        refPage: refPage.current,
-        allPages,
-        totalMed,
         append,
         remove,
         droneId,
