+++ src/components/CatalogMain.js
@@ -1,4 +1,4 @@
-import { React, useContext } from "react";
+import { React, useState, useRef, useContext, useEffect } from "react";
 import FilterAside from "./FilterAside";
 import FilterTop from "./FilterTop";
 import Products from "./Products";
@@ -12,12 +12,97 @@ export default function CatalogMain() {
         behavior: "smooth"
     };
 
+    const [allMedications, setAllMedications] = useState([]);
+    const [availability, setAvailability] = useState(true);
+    const refPage = useRef(0);
+    const [allPages, setAllPages] = useState(0);
+    const [totalMed, setTotalMed] = useState(0);
+
+    const urlFilters = {
+        available: "status=AVAILABLE",
+        sort: {
+            nameAZ: "sort=name,asc",
+            nameZA: "sort=name,desc",
+            weightUp: "sort=weight,asc",
+            weightDown: "sort=weight,desc",
+        }
+    };
+    const [urlSelect, setUrlSelect] = useState(urlFilters.sort.nameAZ);
+    const [urlMedication, setUrlMedication] = useState(refPage.current + "&" + urlFilters.available + "&" + urlSelect);
+
+    useEffect(() => {
+        let ignore = false;
+        fetch("http://localhost:8090/medication?size=6&page=" + urlMedication, {
+            method: 'GET',
+            headers: { 'Content-Type': 'application/json' }
+        })
+            .then(result => result.json())
+            .then(data => {
+                if (!ignore) {
+                    setAllPages(data.totalPages);
+                    setTotalMed(data.totalElements);
+                    updateWithCartInfo(data.content);
+                }
+            });
+        return () => {
+            ignore = true;
+        }
+    }, [urlMedication, context.cartItemsId]);
+
+    const updateWithCartInfo = (medications) => {
+        const updateMedications = medications.map(med => {
+            const isInCart = JSON.parse(localStorage.getItem('cart'))?.find(item => item === med.id);
+            if (isInCart) {
+                return { ...med, status: "UNAVAILABLE" };
+            } else {
+                return { ...med, status: med.status };
+            }
+        });
+        setAllMedications(updateMedications);
+    }
+
+    const toggleAvailable = (value) => {
+        refPage.current = 0;
+        setAvailability(value);
+        setUrlMedication(value
+            ? refPage.current + "&" + urlFilters.available + "&" + urlSelect
+            : refPage.current + "&" + urlSelect);
+    }
+
+    const selectSort = (value) => {
+        refPage.current = 0;
+        for (let key in urlFilters.sort) {
+            if (value === key) {
+                setUrlMedication(availability
+                    ? refPage.current + "&" + urlFilters.available + "&" + urlFilters.sort[value]
+                    : refPage.current + "&" + urlFilters.sort[value]);
+                setUrlSelect(urlFilters.sort[value]);
+            }
+        }
+    }
+
+    const currentPage = (value) => {
+        switch (value) {
+            case "left":
+                refPage.current = refPage.current === 0 ? refPage.current = 0 : refPage.current - 1;
+                break;
+            case "right":
+                refPage.current = refPage.current === allPages - 1
+                    ? allPages - 1 : refPage.current + 1;
+                break;
+            default: refPage.current = value;
+        }
+        setUrlMedication(availability
+            ? refPage.current + "&" + urlFilters.available + "&" + urlSelect
+            : refPage.current + "&" + urlSelect);
+    }
+
     const addPages = () => {
         const arrPage = [];
-        for (let i = 0; i < context.allPages; i++) {
-            arrPage.push(<li key={i} className={i === context.refPage ? "current" : undefined}
+        for (let i = 0; i < allPages; i++) {
+            arrPage.push(<li key={i} className={i === refPage.current ? "current" : undefined}
                 onClick={() => {
-                    context.currentPage(i);
+                    currentPage(i);
                     window.scrollTo(scrollOptions);
                 }}>{i + 1}</li>);
         }
@@ -28,15 +113,15 @@ export default function CatalogMain() {
         <div className="catalog">
             <h1>Catalog</h1>
             <div>
-                <FilterAside />
+                <FilterAside availability={availability} toggleAvailable={toggleAvailable} totalMed={totalMed} />
                 <div className="catalog-filtered">
-                    <FilterTop />
-                    <Products />
+                    <FilterTop availability={availability} toggleAvailable={toggleAvailable} selectSort={selectSort} />
+                    <Products allMedications={allMedications} />
                     <hr />
                     <div className="catalog-pages">
                         <div className="choose-page">
                             <button id="left" onClick={() => {
-                                context.currentPage("left");
+                                currentPage("left");
                                 window.scrollTo(scrollOptions);
                             }}>
                                 <img src="./images/icons/chevron-left.svg" alt="arrow-left" />
@@ -45,7 +130,7 @@ export default function CatalogMain() {
                                 {addPages()}
                             </ul>
                             <button id="right" onClick={() => {
-                                context.currentPage("right");
+                                currentPage("right");
                                 window.scrollTo(scrollOptions);
                             }}>
                                 <img src="./images/icons/chevron-right.svg" alt="arrow-right" />
